25-web_scraping_project /requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
pandas==1.5.3
numpy==1.24.3
scrapy==2.11.0
selenium==4.15.2
lxml==4.9.3
fake-useragent==1.4.0
urllib3==1.26.18
pyyaml==6.0.1
tqdm==4.66.1
html5lib==1.1